SAK-21624
First cut at the Javascript in-browser checking.

Rename the cachetimer for the in-server check to be 

iframe.xframe.cachetime=7200000 

Add new parameter here;

iframe.xframe.loadtime=4000

The amount of time the browser will wait before checking to see if the frame 
load has at least started.

I think the JS needs a bit of tweaking still.



git-svn-id: https://source.sakaiproject.org/svn/web/trunk@120166 66ffb92e-73f9-0310-93c1-f5514f145a0a